Made unit tests.  Broke into modules.
