+++ test.csv
@@ -1,2 +1,3 @@
-﻿a,"Test, this is."
+target, line
+a,"Test, this is."
 b,second line
